+++ packages/assets-controllers/src/token-service.ts
@@ -21,7 +21,7 @@ export const TOKEN_METADATA_NO_SUPPORT_ERROR =
 function getTokensURL(chainId: Hex) {
   const occurrenceFloor = chainId === ChainId['linea-mainnet'] ? 1 : 3;
   // TODO: Either fix this lint violation or explain why it's necessary to ignore.
-  // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
+
   return `${TOKEN_END_POINT_API}/tokens/${convertHexToDecimal(
     chainId,
   )}?occurrenceFloor=${occurrenceFloor}&includeNativeAssets=false&includeTokenFees=false&includeAssetType=false&includeERC20Permit=false&includeStorage=false`;
@@ -36,11 +36,20 @@ function getTokensURL(chainId: Hex) {
  */
 function getTokenMetadataURL(chainId: Hex, tokenAddress: string) {
   // TODO: Either fix this lint violation or explain why it's necessary to ignore.
-  // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
+
   return `${TOKEN_END_POINT_API}/token/${convertHexToDecimal(
     chainId,
   )}?address=${tokenAddress}`;
 }
+
+/**
+ * The sort by field for trending tokens.
+ */
+export type SortTrendingBy =
+  | 'm5_trending'
+  | 'h1_trending'
+  | 'h6_trending'
+  | 'h24_trending';
 
 /**
  * Get the token search URL for the given networks and search query.
@@ -56,6 +65,53 @@ function getTokenSearchURL(chainIds: CaipChainId[], query: string, limit = 10) {
     .map((id) => encodeURIComponent(id))
     .join(',');
   return `${TOKEN_END_POINT_API}/tokens/search?chainIds=${encodedChainIds}&query=${encodedQuery}&limit=${limit}`;
+}
+
+/**
+ * Get the trending tokens URL for the given networks and search query.
+ *
+ * @param chainIds - Array of CAIP format chain IDs (e.g., ['eip155:1', 'eip155:137', 'solana:5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp']).
+ * @param sortBy - The sort by field.
+ * @param minLiquidity - The minimum liquidity.
+ * @param minVolume24hUsd - The minimum volume 24h in USD.
+ * @param maxVolume24hUsd - The maximum volume 24h in USD.
+ * @param minMarketCap - The minimum market cap.
+ * @param maxMarketCap - The maximum market cap.
+ * @returns The trending tokens URL.
+ */
+function getTrendingTokensURL(
+  chainIds: CaipChainId[],
+  sortBy?: SortTrendingBy,
+  minLiquidity?: number,
+  minVolume24hUsd?: number,
+  maxVolume24hUsd?: number,
+  minMarketCap?: number,
+  maxMarketCap?: number,
+) {
+  const encodedChainIds = chainIds
+    .map((id) => encodeURIComponent(id))
+    .join(',');
+  // Add the rest of query params if they are defined
+  const queryParams = new URLSearchParams();
+  if (sortBy) {
+    queryParams.append('sortBy', sortBy);
+  }
+  if (minLiquidity) {
+    queryParams.append('minLiquidity', minLiquidity.toString());
+  }
+  if (minVolume24hUsd) {
+    queryParams.append('minVolume24hUsd', minVolume24hUsd.toString());
+  }
+  if (maxVolume24hUsd) {
+    queryParams.append('maxVolume24hUsd', maxVolume24hUsd.toString());
+  }
+  if (minMarketCap) {
+    queryParams.append('minMarketCap', minMarketCap.toString());
+  }
+  if (maxMarketCap) {
+    queryParams.append('maxMarketCap', maxMarketCap.toString());
+  }
+  return `${TOKEN_END_POINT_API}/v3/tokens/trending?chainIds=${encodedChainIds}&${queryParams.toString()}`;
 }
 
 const tenSecondsInMilliseconds = 10_000;
@@ -135,6 +191,83 @@ export async function searchTokens(
 }
 
 /**
+ * The trending asset type.
+ */
+export type TrendingAsset = {
+  assetId: string;
+  name: string;
+  symbol: string;
+  decimals: number;
+  price: string;
+  aggregatedUsdVolume: number;
+  marketCap: number;
+  priceChangePct?: {
+    m5?: string;
+    m15?: string;
+    m30?: string;
+    h1?: string;
+    h6?: string;
+    h24?: string;
+  };
+  labels?: string[];
+};
+
+/**
+ * Get the trending tokens for the given chains.
+ *
+ * @param options - Options for getting trending tokens.
+ * @param options.chainIds - The chains to get the trending tokens for.
+ * @param options.sortBy - The sort by field.
+ * @param options.minLiquidity - The minimum liquidity.
+ * @param options.minVolume24hUsd - The minimum volume 24h in USD.
+ * @param options.maxVolume24hUsd - The maximum volume 24h in USD.
+ * @param options.minMarketCap - The minimum market cap.
+ * @param options.maxMarketCap - The maximum market cap.
+ * @returns The trending tokens.
+ * @throws Will throw if the request fails.
+ */
+export async function getTrendingTokens({
+  chainIds,
+  sortBy,
+  minLiquidity,
+  minVolume24hUsd,
+  maxVolume24hUsd,
+  minMarketCap,
+  maxMarketCap,
+}: {
+  chainIds: CaipChainId[];
+  sortBy?: SortTrendingBy;
+  minLiquidity?: number;
+  minVolume24hUsd?: number;
+  maxVolume24hUsd?: number;
+  minMarketCap?: number;
+  maxMarketCap?: number;
+}): Promise<TrendingAsset[]> {
+  if (chainIds.length === 0) {
+    console.error('No chains provided');
+    return [];
+  }
+
+  const trendingTokensURL = getTrendingTokensURL(
+    chainIds,
+    sortBy,
+    minLiquidity,
+    minVolume24hUsd,
+    maxVolume24hUsd,
+    minMarketCap,
+    maxMarketCap,
+  );
+
+  try {
+    const result = await handleFetch(trendingTokensURL);
+    return result;
+  } catch (error) {
+    console.error('Trending tokens request failed:', error);
+    return [];
+  }
+}
+
+/**
  * Fetch metadata for the token address provided for a given network. This request is cancellable
  * using the abort signal passed in.
  *
@@ -146,7 +279,7 @@ export async function searchTokens(
  * @returns The token metadata, or `undefined` if the request was either aborted or failed.
  */
 // TODO: Either fix this lint violation or explain why it's necessary to ignore.
-// eslint-disable-next-line @typescript-eslint/naming-convention
+
 export async function fetchTokenMetadata<T>(
   chainId: Hex,
   tokenAddress: string,
@@ -209,7 +342,7 @@ async function parseJsonResponse(apiResponse: Response): Promise<unknown> {
   // api may return errors as json without setting an error http status code
   if (responseObj?.error) {
     // TODO: Either fix this lint violation or explain why it's necessary to ignore.
-    // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
+
     throw new Error(`TokenService Error: ${responseObj.error}`);
   }
   return responseObj;